Changed some testcases and added testcase for 11.9, 14.1, 14.2

+++ misra-c-rules/directory_to_check_all_rules/R_11_6_warn.c
@@ -5,8 +5,8 @@ A cast shall not be performed between pointer to void and an arithmetic type
 #include <stdio.h>
 
 int main(){
-    int x = 2;
+    long x = 2;
     void* p = &x;
-    int r = (int) p;
+    long r = (long) p;
     return 0;
 }
